perf: :arrow_up:  upgrade  vue,vuetify,vuex,vue-router,:racehorse:  complete plugin pass verify, :ar
