+++ nginx.conf.template
@@ -10,7 +10,7 @@ server {
   access_log  /var/log/nginx/host.access.log  main;
   error_log   /var/log/nginx/error.log  error;
 
-  location ~* \.(css|js|png|jpg|jpeg|gif|gz|svg|mp4|ogg|ogv|webm|htc|xml|woff)$ {
+  location ~* ^(?!/api/).*\.(css|js|png|jpg|jpeg|gif|gz|svg|mp4|ogg|ogv|webm|htc|xml|woff)$ {
     root   /usr/src/app/dist/;
     add_header Cache-Control "public, max-age=86400";
     gzip on;
