+++ src/router/dashboard.js
@@ -30,7 +30,6 @@ export const dashboard = [
               show: true,
               rootName: 'dashboard',
               tip: 'dashboard',
-              dependencies: ['kubegems_local'],
             },
           },
         ],
@@ -58,7 +57,6 @@ export const dashboard = [
               show: true,
               rootName: 'dashboard',
               tip: 'project',
-              dependencies: ['kubegems_local'],
             },
           },
         ],
