+++ src/router/workspace.js
@@ -30,7 +30,6 @@ export const workspace = [
               show: true,
               rootName: 'workspace',
               tip: 'environment',
-              dependencies: ['kubegems_local'],
             },
           },
         ],
@@ -72,7 +71,7 @@ export const workspace = [
               show: false,
               rootName: 'workspace',
               tip: 'app',
-              dependencies: ['kubegems_local', 'argocd'],
+              dependencies: ['argocd'],
             },
           },
           {
@@ -86,7 +85,7 @@ export const workspace = [
               show: false,
               rootName: 'workspace',
               tip: 'deploy',
-              dependencies: ['kubegems_local', 'argocd', 'argo_rollouts'],
+              dependencies: ['argocd', 'argo_rollouts'],
             },
           },
         ],
