+++ src/views/microservice/components/ResourceInfo/MonitorCard/index.vue
@@ -250,8 +250,8 @@
 
 <script>
 import { mapState } from 'vuex'
-import { matrix, vector } from '@/api'
 import WorkloadInfo from './WorkloadInfo'
+import BasePermission from '@/mixins/permission'
 import {
   ISTIO_WORKLOAD_QPS_PROMQL,
   ISTIO_WORKLOAD_REQUEST_LAST_24H_PROMQL,
@@ -266,6 +266,7 @@ import {
 export default {
   name: 'MonitorCard',
   components: { WorkloadInfo },
+  mixins: [BasePermission],
   data() {
     return {
       requests: [],
@@ -316,7 +317,7 @@ export default {
       this.istioWorkloadNetworkEgressSum(true)
     },
     async istioWorkloadRequestCount(noprocess = false) {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.$route.query.cluster,
         Object.assign(this.params, {
           query: ISTIO_WORKLOAD_QPS_PROMQL
@@ -331,7 +332,7 @@ export default {
       }
     },
     async istioWorkloadRequestCountNow(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_WORKLOAD_QPS_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -339,12 +340,12 @@ export default {
       })
       data.forEach((d) => {
         this.requestsNow = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
     async istioWorkloadRequestSumOverTime(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_WORKLOAD_REQUEST_LAST_24H_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -352,13 +353,13 @@ export default {
       })
       data.forEach((d) => {
         this.requestsSumOverTime = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
 
     async istioWorkloadResponseDurationSeconds(noprocess = false) {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.$route.query.cluster,
         Object.assign(this.params, {
           query: ISTIO_WORKLOAD_RESPONSE_DURATION_SECONDS_P95_PROMQL
@@ -373,7 +374,7 @@ export default {
       }
     },
     async istioWorkloadResponseDurationSecondsP95(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_WORKLOAD_RESPONSE_DURATION_SECONDS_P95_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -381,12 +382,12 @@ export default {
       })
       data.forEach((d) => {
         this.avgresponsetimeP95 = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] * 1000,
+          isNaN(d?.value[1]) ? 0 : d.value[1] * 1000,
         ).toFixed(2)
       })
     },
     async istioWorkloadResponseDurationSecondsP99(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_WORKLOAD_RESPONSE_DURATION_SECONDS_P99_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -394,13 +395,13 @@ export default {
       })
       data.forEach((d) => {
         this.avgresponsetimeP99 = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] * 1000,
+          isNaN(d?.value[1]) ? 0 : d.value[1] * 1000,
         ).toFixed(2)
       })
     },
 
     async istioWorkloadErrResponseCount(noprocess = false) {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.$route.query.cluster,
         Object.assign(this.params, {
           query: ISTIO_WORKLOAD_ERR_QPS_PROMQL
@@ -415,7 +416,7 @@ export default {
       }
     },
     async istioWorkloadErrResponseSum(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_WORKLOAD_ERR_QPS_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -423,12 +424,12 @@ export default {
       })
       data.forEach((d) => {
         this.errrequestsSum = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
     async istioWorkloadErrResponseSumOverTime(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_WORKLOAD_ERR_REQUEST_LAST_24H_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -436,13 +437,13 @@ export default {
       })
       data.forEach((d) => {
         this.errrequestsSumOverTime = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
 
     async istioWorkloadNetwork(noprocess = false) {
-      const data1 = await matrix(
+      const data1 = await this.m_permission_matrix(
         this.$route.query.cluster,
         Object.assign(this.params, {
           query: ISTIO_INGRESS_NETWORK_PROMQL
@@ -452,7 +453,7 @@ export default {
         }),
       )
       if (data1 && data1.length > 0) data1[0].metric['name'] = 'ingress流量'
-      const data2 = await matrix(
+      const data2 = await this.m_permission_matrix(
         this.$route.query.cluster,
         Object.assign(this.params, {
           query: ISTIO_EGRESS_NETWORK_PROMQL
@@ -469,7 +470,7 @@ export default {
       this.network = data
     },
     async istioWorkloadNetworkIngressSum(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_INGRESS_NETWORK_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -477,12 +478,12 @@ export default {
       })
       data.forEach((d) => {
         this.networkIngress = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] / 1024,
+          isNaN(d?.value[1]) ? 0 : d.value[1] / 1024,
         ).toFixed(2)
       })
     },
     async istioWorkloadNetworkEgressSum(noprocess = false) {
-      const data = await vector(this.$route.query.cluster, {
+      const data = await this.m_permission_vector(this.$route.query.cluster, {
         query: ISTIO_EGRESS_NETWORK_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -490,7 +491,7 @@ export default {
       })
       data.forEach((d) => {
         this.networkEgress = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] / 1024,
+          isNaN(d?.value[1]) ? 0 : d.value[1] / 1024,
         ).toFixed(2)
       })
     },
