+++ src/views/observe/overview/components/OverviewList/K8sEvents/MessageBarChart.vue
@@ -62,8 +62,8 @@ export default {
         enabled: false,
       },
       stroke: {
-        curve: 'straight',
-        width: 1,
+        curve: 'smooth',
+        width: 0,
       },
       grid: {
         borderColor: 'rgba(0, 0, 0, .3)',
@@ -110,6 +110,7 @@ export default {
       },
       noData: {
         text: '暂无数据',
+        offsetY: -28,
       },
     }
     return {
