+++ src/views/resource/cluster/components/ResourceChart.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-row>
+  <v-row class="mt-1">
     <v-col cols="3">
       <v-card>
         <v-card-text>
