+++ src/views/resource/dashboard/components/ResourceList/index.vue
@@ -262,3 +262,4 @@ export default {
   },
 }
 </script>
+
