+++ src/views/resource/deploy/components/DeployControlCenter/DeployCard.vue
@@ -103,8 +103,8 @@
 </template>
 
 <script>
-import { vector } from '@/api'
 import BaseResource from '@/mixins/resource'
+import BasePermission from '@/mixins/permission'
 import {
   ISTIO_WORKLOAD_QPS_PROMQL,
   ISTIO_WORKLOAD_REQUEST_LAST_24H_PROMQL,
@@ -118,7 +118,7 @@ import {
 
 export default {
   name: 'DeployCard',
-  mixins: [BaseResource],
+  mixins: [BaseResource, BasePermission],
   props: {
     status: {
       type: Object,
@@ -167,7 +167,7 @@ export default {
       this.istioWorkloadNetworkEgressSum(true)
     },
     async istioWorkloadRequestCountNow(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_WORKLOAD_QPS_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -175,12 +175,12 @@ export default {
       })
       data.forEach((d) => {
         this.requestsNow = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
     async istioWorkloadRequestSumOverTime(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_WORKLOAD_REQUEST_LAST_24H_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -188,13 +188,13 @@ export default {
       })
       data.forEach((d) => {
         this.requestsSumOverTime = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
 
     async istioWorkloadResponseDurationSecondsP99(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_WORKLOAD_RESPONSE_DURATION_SECONDS_P99_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -202,12 +202,12 @@ export default {
       })
       data.forEach((d) => {
         this.avgresponsetimeP99 = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] * 1000,
+          isNaN(d?.value[1]) ? 0 : d.value[1] * 1000,
         ).toFixed(2)
       })
     },
     async istioWorkloadResponseDurationSecondsP95(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_WORKLOAD_RESPONSE_DURATION_SECONDS_P95_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -215,13 +215,13 @@ export default {
       })
       data.forEach((d) => {
         this.avgresponsetimeP95 = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] * 1000,
+          isNaN(d?.value[1]) ? 0 : d.value[1] * 1000,
         ).toFixed(2)
       })
     },
 
     async istioWorkloadErrResponseSum(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_WORKLOAD_ERR_QPS_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -229,12 +229,12 @@ export default {
       })
       data.forEach((d) => {
         this.errrequestsSum = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
     async istioWorkloadErrResponseSumOverTime(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_WORKLOAD_ERR_REQUEST_LAST_24H_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -242,13 +242,13 @@ export default {
       })
       data.forEach((d) => {
         this.errrequestsSumOverTime = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1],
+          isNaN(d?.value[1]) ? 0 : d.value[1],
         ).toFixed(2)
       })
     },
 
     async istioWorkloadNetworkIngressSum(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_INGRESS_NETWORK_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -256,12 +256,12 @@ export default {
       })
       data.forEach((d) => {
         this.networkIngress = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] / 1024,
+          isNaN(d?.value[1]) ? 0 : d.value[1] / 1024,
         ).toFixed(2)
       })
     },
     async istioWorkloadNetworkEgressSum(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: ISTIO_EGRESS_NETWORK_PROMQL
           .replaceAll('$1', this.$route.params.name)
           .replaceAll('$2', this.$route.query.namespace),
@@ -269,7 +269,7 @@ export default {
       })
       data.forEach((d) => {
         this.networkEgress = parseFloat(
-          isNaN(d.value[1]) ? 0 : d.value[1] / 1024,
+          isNaN(d?.value[1]) ? 0 : d.value[1] / 1024,
         ).toFixed(2)
       })
     },
