+++ src/views/resource/environment/components/ResourceMonitor.vue
@@ -24,6 +24,7 @@
             type="cpu"
             :label-show="false"
             label="environment"
+            :no-data-offset-y="-23"
           />
         </v-col>
         <v-col cols="6">
@@ -34,6 +35,7 @@
             type="memory"
             :label-show="false"
             label="environment"
+            :no-data-offset-y="-23"
           />
         </v-col>
       </v-row>
@@ -46,6 +48,7 @@
             type="network"
             :label-show="false"
             label="environment"
+            :no-data-offset-y="-23"
           />
         </v-col>
         <v-col cols="6">
@@ -56,6 +59,7 @@
             type="network"
             :label-show="false"
             label="environment"
+            :no-data-offset-y="-23"
           />
         </v-col>
       </v-row>
@@ -65,8 +69,8 @@
 
 <script>
 import { mapGetters, mapState } from 'vuex'
-import { matrix } from '@/api'
 import BaseResource from '@/mixins/resource'
+import BasePermission from '@/mixins/permission'
 import {
   ENVIRONMENT_CPU_USAGE_PROMQL,
   ENVIRONMENT_MEMORY_USAGE_PROMQL,
@@ -76,7 +80,7 @@ import {
 
 export default {
   name: 'ResourceMonitor',
-  mixins: [BaseResource],
+  mixins: [BaseResource, BasePermission],
   props: {
     ready: {
       type: Boolean,
@@ -137,7 +141,7 @@ export default {
       this.environmentNetworkOut()
     },
     async environmentCPUUsage() {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.ThisCluster,
         Object.assign(this.params, {
           query: ENVIRONMENT_CPU_USAGE_PROMQL.replaceAll(
@@ -150,7 +154,7 @@ export default {
       if (data) this.cpu = data
     },
     async environmentMemoryUsage() {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.ThisCluster,
         Object.assign(this.params, {
           query: ENVIRONMENT_MEMORY_USAGE_PROMQL.replaceAll(
@@ -163,7 +167,7 @@ export default {
       if (data) this.memory = data
     },
     async environmentNetworkIn() {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.ThisCluster,
         Object.assign(this.params, {
           query: ENVIRONMENT_NETWORK_IN_PROMQL.replaceAll(
@@ -176,7 +180,7 @@ export default {
       if (data) this.networkin = data
     },
     async environmentNetworkOut() {
-      const data = await matrix(
+      const data = await this.m_permission_matrix(
         this.ThisCluster,
         Object.assign(this.params, {
           query: ENVIRONMENT_NETWORK_OUT_PROMQL.replaceAll(
