+++ src/views/resource/node/index.vue
@@ -269,10 +269,11 @@
 
 <script>
 import { mapState } from 'vuex'
-import { getNodeList, patchCordonNode, vector } from '@/api'
+import { getNodeList, patchCordonNode } from '@/api'
 import GpuScheduleForm from './components/GpuScheduleForm'
 import BaseFilter from '@/mixins/base_filter'
 import BaseResource from '@/mixins/resource'
+import BasePermission from '@/mixins/permission'
 import {
   NODE_LOAD_PROMQL,
   NODE_ALL_CPU_USAGE_PROMQL,
@@ -286,7 +287,7 @@ export default {
   components: {
     GpuScheduleForm,
   },
-  mixins: [BaseFilter, BaseResource],
+  mixins: [BaseFilter, BaseResource, BasePermission],
   data: () => ({
     items: [],
     headers: [
@@ -358,7 +359,7 @@ export default {
       }, 1000 * 60)
     },
     async nodeLoad5(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: NODE_LOAD_PROMQL,
         noprocessing: noprocess,
       })
@@ -368,13 +369,13 @@ export default {
         })
         if (index > -1) {
           const item = this.items[index]
-          item.load = parseFloat(d.value[1]).toFixed(1)
+          item.load = parseFloat(d?.value[1]).toFixed(1)
           this.$set(this.items, index, item)
         }
       })
     },
     async nodeCPUUsage(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: NODE_ALL_CPU_USAGE_PROMQL,
         noprocessing: noprocess,
       })
@@ -385,16 +386,16 @@ export default {
         if (index > -1) {
           const item = this.items[index]
           item.cpu = (
-            (parseFloat(d.value[1]) * item.status.capacity.cpu) /
+            (parseFloat(d?.value[1]) * item.status.capacity.cpu) /
             100
           ).toFixed(1)
-          item.cpuPercentage = parseFloat(d.value[1]).toFixed(1)
+          item.cpuPercentage = parseFloat(d?.value[1]).toFixed(1)
           this.$set(this.items, index, item)
         }
       })
     },
     async nodeMemoryUsage(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: NODE_ALL_MEMORY_USAGE_PROMQL,
         noprocessing: noprocess,
       })
@@ -405,19 +406,19 @@ export default {
         if (index > -1) {
           const item = this.items[index]
           item.memory = (
-            (parseFloat(d.value[1]) *
+            (parseFloat(d?.value[1]) *
               sizeOfStorage(item.status.capacity.memory, 'Ki')) /
             1024 /
             1024 /
             100
           ).toFixed(1)
-          item.memoryPercentage = parseFloat(d.value[1]).toFixed(1)
+          item.memoryPercentage = parseFloat(d?.value[1]).toFixed(1)
           this.$set(this.items, index, item)
         }
       })
     },
     async nodePodCount(noprocess = false) {
-      const data = await vector(this.ThisCluster, {
+      const data = await this.m_permission_vector(this.ThisCluster, {
         query: NODE_POD_RUNNING_COUNT_PROMQL,
         noprocessing: noprocess,
       })
@@ -427,9 +428,9 @@ export default {
         })
         if (index > -1) {
           const item = this.items[index]
-          item.podcount = parseInt(d.value[1])
+          item.podcount = parseInt(d?.value[1])
           item.podPercentage = (
-            (parseInt(d.value[1]) / item.status.capacity.pods) *
+            (parseInt(d?.value[1]) / item.status.capacity.pods) *
             100
           ).toFixed(1)
           this.$set(this.items, index, item)
