+++ src/views/tenant/tenant/components/ResourceList/ResourceChart.vue
